feat(burger-menu): Integrated sign-up/sign-in pages into the menu
The burger menu now includes links for registering and signing in when there is no current user, or
a link to logout when there is a logged in user. It also displays the current user's email in order
to ensure they are signed in and in the right account. Routes were also moved into another file to
modularize the code further and make for easier debugging in the future.