feat(cronofy): Create Settings and External calendar views, and implement user-side OAuth for Cronof
Created new, mostly empty Settings View. Also, user can now authorize Commulist with Cronofy. When
they do so, Cronofy will ask for the credentials for any of the supported calendars. Afterward, per
the OAuth2 spec, Cronofy redirects the user to a URI of our choice with a one-time use code in a
query param that Commulist can exchange with Cronofy for an auth token. We need to store that auth
token in Meteor.user(), set up subscriptions to Cronofy, and we'll be cookin'.